src/test/javascript/spec/app/entities/services-ride-fixxer/services-ride-fixxer.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { ServicesRideFixxerService } from 'app/entities/services-ride-fixxer/services-ride-fixxer.service';
import { IServicesRideFixxer, ServicesRideFixxer } from 'app/shared/model/services-ride-fixxer.model';

describe('Service Tests', () => {
    describe('ServicesRideFixxer Service', () => {
        let injector: TestBed;
        let service: ServicesRideFixxerService;
        let httpMock: HttpTestingController;
        let elemDefault: IServicesRideFixxer;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(ServicesRideFixxerService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new ServicesRideFixxer(0, 'AAAAAAA', 0, currentDate);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        timeAprox: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a ServicesRideFixxer', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        timeAprox: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        timeAprox: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new ServicesRideFixxer(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a ServicesRideFixxer', async () => {
                const returnedFromService = Object.assign(
                    {
                        name: 'BBBBBB',
                        price: 1,
                        timeAprox: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        timeAprox: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of ServicesRideFixxer', async () => {
                const returnedFromService = Object.assign(
                    {
                        name: 'BBBBBB',
                        price: 1,
                        timeAprox: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        timeAprox: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a ServicesRideFixxer', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
